+++ esdm/esdm_drng_mgr.h
@@ -97,9 +97,6 @@ int esdm_drng_mgr_initialize(void);
 void esdm_drng_mgr_finalize(void);
 bool esdm_get_available(void);
 void esdm_drng_reset(struct esdm_drng *drng);
-void esdm_drng_inject(struct esdm_drng *drng, const uint8_t *inbuf,
-		      size_t inbuflen, bool fully_seeded,
-		      const char *drng_type);
 void esdm_drng_seed_work(void);
 void esdm_try_fully_seeded(void);
 void esdm_force_fully_seeded(void);